Better error handling on administration page
Now when the request fails the error message is based
on the status code returned by the API. This will
help handling unexpected messages, and create better bug
reports.

The layout of the page has also been reworked in order
to make the alerts visually related to the form from
they were issued. In better words alerts are displayed right
under the form which raised the error, instead of the middle
of the page.

Fix https://github.com/redhat-cip/dci-ui/issues/1

Change-Id: I049beaa56950b3915f37402935c9005071dec086