Better workflow for revision generation
Add error handling during  generation of the revision number in
config.json.

Add git as a dependency for generating the revision number at
package build.

The config.json is now generated at build from the config.json.tplt
file. This avoid confusions in the config.json file which was
versionning the version of the front.

Change build_rpm for fetching HEAD instead of master, this will avoid
to miss new directories.

Change-Id: I924596574ffe1ad2407c7719ada4a486b6f13a67