Fix display the first jobstates files
Sometimes jobstates files were not displayed correctly, i.e:
the last jobstate with files is opened and displays files content

This was due to the fact that we track promises and the first one
which had files in it told the jobstate object to display files.
But, due to the asynchronous way of retrieving files, if the last
jobstate had a huge amount of files, some callback could resolve
first. This leading to display the wrong files.

This patch introduce a utility function which resolve promises in
a defined order. This function is used in order to display correctly
the last jobstate with files.

Change-Id: Ia2d13cd3c1ab0301700b8e092f34d7b034cac6ad