Add a messages services
This service allows the user to send messages to the UI,
those messages will be displayed in a dedicated part of the
interface.
Messages can be closed by hand or they will fade after a
configurable time.
This patch removes all the use of custom alerts mecanisms
in the code.
This patch also fix the 404 errors when trying to reach
an unexisting job, and redirect to the index.

Change-Id: I430c0a5995e278c0bfaf77e0f2a9363bb9050a23