Update patch set 3
Patch Set 3: Code-Review-1

I realize I was not clear enough in the story description.

What you did is a JSON viewer, which nice. In this case we should stcik the JSON standard and use instead: application/json.

For the tripleo-stack-dump output, I think the data structure is huge and we should just show up some keys. For instance: jq .stacks.overcloud.outputs.parameters tripleo-stack-dump-sample.json

Later, we will be able to show more informations. For example, the list of resources with their status.

Can we use application/tripleo-stack-dump for tripleo-stack-dump? This is what I already use for DELL.

Patch-set: 3
Label: Code-Review=-1