Update patch set 3
Patch Set 3:

In a lot of cases, we do not manage to get an overcloud running, in this case, the test are just useless and we just get the 'No JUnit result uploaded.' message.

So we can reorder the tab, but we can also generate a junit file and upload it:


    Provisioning: success/failure
    Undercloud deployment: success/failure
    Overcloud deployment: success/failure

This way, we will show up the deployment itself as a test and get the nice Junit results.

So, Yanis, I'm find with your position.


PS:I'm not sure Jenkins is a good example here, 99% of the time, when I reach the job page, it's to click on the 'Show log' link.

Patch-set: 3