create_at: Create a second formatted field for display.
In its current state, the code replace the content of the created_at
value by a formatted - only for display - value.

We should keep the original version of the created_at field as returned
by the API in the object, and create a new created_at_formatted field
that will be used just for display. The same pattern applies to
updated_at.

On of the issue, is for example sorting. Sorting by date is currently
broken because we replace the value by the formatted value.

Currently it sorts:

2016-10-06
2016-09-30
2016-09-23
2016-09-12

The following way (alphabecitcally based on the first letter of the
day):

Thursday, October 6, 2016 9:07 PM
Monday, September 12, 2016 4:41 PM
Friday, September 30, 2016 5:31 PM
Friday, September 23, 2016 8:26 PM

Change-Id: Iee89b1c364de42f0a60bb40f275cd2e775142fd8