Update patch set 3
Patch Set 3: Code-Review-1

You need to rebase this onto master

Could we limit this to super admin first?

The fingerprint doesn't make sense right now for a user.
My feeling is that we start to create regex in administration.
Then starts adding automatic issues to jobs.
At this moment we will validate that fingerprint works, and can introduce this in the interface for all admin partners.

Patch-set: 3