job: only show up real files in the files page
There is two type of files:
- those attached to a jobstate, that are actually just the outpout of an
  Ansible step
- and the log files that are attached to the job by the agents.

The first case used to be hidden by default because their job_id was
NULL. Since a couple of months, all the files attached to a job have
their job_id set. As a consequence, we can now see those Ansible task
outputs in the job tab.

This commit intend to reverse this behavior and switch back to the
previous situation.

Change-Id: Idb6a413246486bdccc50ecb5131cf2678849f242