Fix warning in userForm and apply destructuring to every component
The issue was: React does not recognize the `showModalButton` prop on a DOM element.
This is because in UserForm we apply non destructured element to a node.
To avoid this issue we should destructure every props used in the component or use this.props everywhere.
We should avoid to mix both.

We now use destructuring props everywhere.

Change-Id: Iafacb504d06216fb6f3fa518c81546174746bb5c