Create and use configProvider and authProvider
Use react context provider instead of redux for the config and auth mechanism.
Allow us to create useAuth() and useConfig() hooks.

Also use location from react router if not connected.
This allow us to redirect the user where he wants to go after a successful login:
 * Save location in oidc state object for sso mechanism
 * Save location in react router state for basic auth mechanism

We should remove the currentUser in redux store and use identity from authProvider everywhere.
But the patch is so big (my excuses) that I will do it in another patch.

Change-Id: Ide003381e22585d4c8c328a61daffdab1e3e05ad