Remove dynamic config.json
Problem: I would like to speed up the ui startup, simplify the code and use best practice for the frontend.
Today when the user open https://www.distributed-ci.io/ we have :

 * The javascript is downloaded
 * Javascript request https://www.distributed-ci.io/config.json to get the configuration (api url and sso configuration)
 * We save this information in a javascript store (global variable)
 * Everytime we would like to do a request we get the API url in the store (see state.config.apiURL)

This config doesn't change very often. Best practice is to set those config variables at compile time and produce a web app with hardcoded value. We use env variables to do this.

This patch remove the need of the config file.

Change-Id: I3824afffbadb0c58d8db334949eac3d776b543f7