Improve files page
When a user use the interface, the link on the file name in the files list is disturbing.
A click redirects the user to a empty page with the file content.
I change the behavior by opening the content of the file in a popup.
We keep the new page behavior by adding a link button in the actions.
This patch also fix humanFileSize returning NaN in cas of a file size to 0.
Non regression test is present.
Also we don't filter modal view with file content. User will see binary content.

Change-Id: I9c8e4d419c3c03ccb4d9691e340a3e06f081d2b6