Add job state pipeline
A user must scroll directly to the bottom of a page to go to the last task of a job state.
This review is made to represent the job states in the form of a pipeline, to allow the user to click directly on a link to go to the desired job state.

Change-Id: If81352fca39a139855663989bbb0eff94a2ba998