Stop filtering by default jobs
In a recent patch I introduce a automatic job filter by team id.
Some users raised the issue that it's counter intuitive.
This patch revert the old patch.

We also save filters in localStorage if the user change the filters.
We remove the localStorage key if the user clear filters.

We fix also an issue where tags:tag1a:tag1b where not parsed correctly.
The tag is 'tag1a:tag1b'

Change-Id: If37a866753f61b3e6b45618e17c75a420c562e00