Improve navigation between job details pages
Job details pages are not handling correctly the navigation history.
If you click on tests then files tab in the job details page, and hit previous on the navigator, then the page return to the previous page, not the tests page.
This is fixed now using Outlet from react-router v6

Change-Id: I617e31cd352ab1ca2b84c364a6847f620eb87e14