Ignore search params change in history
Some parameters are set by default, like page number and elements per pages in the jobs page.
When the user hit the previous page button on his browser, the navigation is not good.
It displayed the actual page without search params instead of the previous one.

This patch use replace: true in react router navigation point when needed.

Change-Id: I0b2e36de744353d4022fff78254a243106b9411c