Display tests summary in jobs table view instead of all tests
In jobs table view, columns are small, and we destroy layout when we add tests because it displays all the tests names and results.
This patch changes this and display only the sum of all tests. To see all tests, the user now have to see the detail of a job.

Change-Id: I1ae9ade78b80b3b493e7ea128f7b3527b0a398c7