Clarify canonical project name and name on the component page
Name for canonical project name is confusing for the partners.
They use the name in the agents to query explicitly some component.
Event if canonical project name doesn't make sense right now, we are going to display it until the component migration occurs.

Use this patch to also display the released_at date.

Change-Id: If9ade7bb9443149e995c86ae8f85e5894ad91e53