Add navigation information in the title
When a user navigate on DCI, he only sees 'Distributed CI' in the navigation list in his browser.
This page change the title of the page when he visit it.
Like that the user can jump in a previous page without having to count mentally to find the good page.

Change-Id: I12471b1801eb28f2ef9cee8ac242dae217967f86