Improve error management in job context
Today when we visit a job detail page and the job is not accessible, we see a endless loading state.
This patch fix the issue and display the error message returned by the backend.

We also improve the feeling of speed by introducing a job skeleton in the job detail page.

Change-Id: I07eb292369f6431cd205c8f4c3d45a780e903557